+++ test/local-svf-to-gltf.js
@@ -9,8 +9,9 @@ const { SvfReader, GltfWriter } = require('..');
 async function run (filepath, outputDir) {
     const reader = await SvfReader.FromFileSystem(filepath);
     const svf = await reader.read();
-    const writer = new GltfWriter();
-    writer.write(svf, outputDir);
+    const writer = new GltfWriter(outputDir);
+    writer.write(svf);
+    writer.close();
 }
 
 run(process.argv[2], process.argv[3]);